Factor our ns exemptions to make them consumable by other projects
